requirements.txt
python-dotenv==1.0.0
langchain-core==0.1.53
langchain-google-genai==0.0.11
google-generativeai==0.3.2